Add a build-condition for disabling building documentation
The optional documentation distributed within doc subpackage requires
doxygen and graphviz at build time. That is an obstacle when building
jsoncpp for modular Fedora when the dependecies are not yet available.

This patch introduces a build-time condition that allows to disable
generating the documentation while building in normal Fedora is
unchanged.